feat: complete architectural migration to clean multi-project structure
## Major Changes:
- **New WitcherCore project**: Clean business logic layer with Models, Services, Configuration
- **New Shared project**: Common types and utilities
- **New DataLayer project**: Database access layer with repositories
- **Added .github/copilot-instructions.md**: Comprehensive AI coding guidelines with PowerShell environment
- **Started save file analysis buildout**: MetadataExtractor and database integration foundation

## Architecture Migration:
- Moved WitcherSaveFileService from frontend/Services to WitcherCore/Services
- Moved GameSaveExtensions, MetadataExtractor, SavePathResolver to WitcherCore
- Created proper Models namespace in WitcherCore (GameKey, WitcherSaveFile, etc.)
- Updated all namespaces from WitcherSmartSaveManager.* to WitcherCore.*
- Removed legacy Services project dependencies

## Save File Analysis:
- Added MetadataExtractor service for parsing save file contents
- Created database foundation with SaveFileRepository and ISaveFileRepository
- Set up SaveFileMetadata model for structured data storage
- Added SQL initialization scripts for database schema

## Test Fixes:
- Updated all tests for required WitcherSaveFile properties (FileName, ModifiedTimeIso)
- Fixed namespace imports in test files
- Removed obsolete SaveFileAnalysisServiceTests (service no longer exists)
- All 30 tests now pass successfully

## Configuration:
- Updated solution file with new project structure
- Added proper project references and dependencies
- Updated .gitignore for new project artifacts
- Enhanced CI/CD workflows for multi-project build

## Code Quality:
- Applied nullable reference type best practices with required properties
- Enforced one-class-per-file organization
- Implemented clean separation of concerns
- Established PowerShell-first development environment

Closes architectural migration phase - clean MVVM structure achieved with save analysis foundation!
