+++ .github/workflows/pester-tests.yml
@@ -24,14 +24,7 @@ jobs:
 
     steps:
     - name: Checkout code
-      uses: actions/checkout@v3
-
-    - name: Set up PowerShell
-      uses: actions/setup-powershell@v2
-
-    - name: Install Pester
-      run: |
-        Install-Module -Name Pester -Force -Scope CurrentUser
+      uses: actions/checkout@v3    
 
     - name: Run Pester Tests
       run: |
