+++ Shared/Models/LanguageResource.cs
@@ -1,0 +1,8 @@
+namespace WitcherCore.Models
+{
+    public class LanguageResource
+    {
+        public string Key { get; set; }
+        public string Value { get; set; }
+    }
+}
