+++ WitcherCore/Data/ISaveFileRepository.cs
@@ -1,0 +1,9 @@
+using WitcherCore.Models;
+
+namespace WitcherCore.Data
+{
+    public interface ISaveFileRepository
+    {
+        void InsertMetadata(SaveFileMetadata metadata);
+    }
+}
