+++ WitcherCore/Models/GameKey.cs
@@ -1,0 +1,9 @@
+namespace WitcherCore.Models
+{
+    public enum GameKey
+    {
+        Witcher1,
+        Witcher2,
+        Witcher3
+    }
+}
