+++ WitcherCore/Models/InstallSource.cs
@@ -4,15 +4,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace WitcherSmartSaveManager.Models
+namespace WitcherCore.Models
 {
-    public enum GameKey
-    {
-        Witcher1,
-        Witcher2,
-        Witcher3
-    }
-
     public enum InstallSource
     {
         Steam,
