+++ WitcherSmartSaveManager/ViewModels/MainViewModel.cs
@@ -1,0 +1,1 @@
+// ...existing code from frontend/ViewModels/MainViewModel.cs will be moved here...
