+++ WitcherSmartSaveManager/Views/MainView.xaml
@@ -1,0 +1,1 @@
+<!-- ...existing XAML code for MainView.xaml will be moved here... -->
