+++ frontend/WitcherSmartSaveManager.csproj
@@ -41,4 +41,9 @@
 	<Resource Include="./Views/Assets/icon_wolf_save.ico" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\WitcherCore\WitcherCore.csproj" />
+    <ProjectReference Include="..\Shared\Shared.csproj" />
+  </ItemGroup>
+
 </Project>
